+++ src/main/java/com/amotassic/dabaosword/effect/BingliangEffect.java
@@ -1,6 +1,6 @@
 package com.amotassic.dabaosword.effect;
 
-import com.amotassic.dabaosword.item.ModItems;
+import com.amotassic.dabaosword.items.ModItems;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
