+++ src/main/java/com/amotassic/dabaosword/enchantment/RageNatureEnchantment.java
@@ -1,7 +1,7 @@
 package com.amotassic.dabaosword.enchantment;
 
 import com.amotassic.dabaosword.Sounds;
-import com.amotassic.dabaosword.item.ModItems;
+import com.amotassic.dabaosword.items.ModItems;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentTarget;
 import net.minecraft.entity.Entity;
