+++ src/main/java/com/amotassic/dabaosword/event/AttackEntityHandler.java
@@ -1,6 +1,6 @@
 package com.amotassic.dabaosword.event;
 
-import com.amotassic.dabaosword.item.ModItems;
+import com.amotassic.dabaosword.items.ModItems;
 import net.fabricmc.fabric.api.event.player.AttackEntityCallback;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.mob.HostileEntity;
