+++ src/main/resources/assets/dabaosword/models/item/huoshaolianying.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "dabaosword:item/huoshaolianying"
+  }
+}
